Fix transducer sequence arity contracts
Transduction should only be attempted on sequences that produce one value at a time. Needs to be checked because it's easy to accidentally use a hash where you should use (in-hash-entries hash), and without checks that leads to hard to debug error messages in the guts of the transducer system.