Rename range-bound-type enum to bound-type
This is a breaking change. This is needed so that define-enum-type can bind the enum name to static information about the enum. This should only affect clients who referred to the range-bound-type? predicate, which is unlikely outside of contracts. Hopefully this doesn't break anyone important - if it did and you're reading this, please file an issue in Rebellion explaining the breakage.