Soft-deprecate reduce and reduce-all
Having three different ways of using reducers is pretty weird, especially when two of them don't support transducers at all and take arguments in the opposite order of transduce. Since transduce is the main API for rebellion/streaming, we discourage the other two forms and rewrite the documentation examples in terms of transduce.