Make `comparator-reverse` an involution
By storing a pre-reversed form of the comparison function in the comparator struct, we can make `comparator-reverse` just swap some fields around instead of allocating a new procedure. This makes `comparator-reverse` an involution, meaning that `(comparator-reverse (comparator-reverse cmp))` is `equal?` to `cmp`. This property is important for sorted collections, since reversing an immutable sorted collection twice should produce a collection that's equal to the original.