Test sorted map key and entry searching
This was mostly already covered by the tests of the key and entry set views, but not all implementations of sorted maps use the same code paths for `(sorted-map-least-key m)` and `(sorted-set-least-element (sorted-map-keys m))`. Array-backed sorted maps, for example, switch wholesale to the array-backed sorted set implementation instead of returning a view because that can be done without copying.