Fix a transducer composition bug.
This stream pipeline currently errors:

```scheme
(transduce (in-range 0 10) (sorting) (mapping values) #:into (into-list))
```

This is due to a bug in the logic for composing two transducers together. When the upstream transducer is finished, and the downstream transducer emits a value, that emission should be changed to a half-closed emission. The current logic doesn't do that and results in a contract error. This commit fixes the issue and adds a test.